+++ crate/input_ir_rt/src/lib.rs
@@ -8,6 +8,8 @@ pub(crate) use crate::{
     taffy_node_absolute_coordinates_calculator::TaffyNodeAbsoluteCoordinatesCalculator,
 };
 pub use crate::{
+    diagram_generate_error::DiagramGenerateError,
+    diagram_generator::DiagramGenerator,
     edge_face_assigner::EdgeFaceAssigner,
     edge_id_generator::EdgeIdGenerator,
     input_diagram_merger::InputDiagramMerger,
@@ -32,6 +34,8 @@ const NOTO_SANS_MONO_TTF: &[u8] =
     include_bytes!("../fonts/noto_sans_mono/NotoSansMono-Regular.ttf");
 
 mod absolute_coordinates;
+mod diagram_generate_error;
+mod diagram_generator;
 mod edge_face_assigner;
 mod edge_id_generator;
 mod input_diagram_merger;