+++ crate/input_model/src/lib.rs
@@ -3,7 +3,7 @@
 //! The diagram input model is hand written, as an OpenAPI spec doesn't support
 //! modelling certain data structures such as a Map with a particular key type.
 
-pub use crate::input_diagram::InputDiagram;
+pub use crate::{diagram_focus::DiagramFocus, input_diagram::InputDiagram};
 
 pub mod edge;
 pub mod entity;
@@ -12,4 +12,5 @@ pub mod tag;
 pub mod theme;
 pub mod thing;
 
+mod diagram_focus;
 mod input_diagram;